Corona: Add WebRequestParser class.
This is the advent of the new style request handling in Lunr.
The existing Request class will be split in parsing and reading
parts, where the reading parts stay as Request class, and the
parsing parts move into dedicated RequestParser classes.

The goal is to make the reading rather generic, so extending
it should only be necessary in corner cases, and request data
parsing can be customized by implementing the RequestParserInterface,
or extending an existing RequestParser class.

Compliance can be tested by using the RequestParserRequestTestTrait.