Corona: Update Request class to depend on request parsers.
This removes the internal request parsing and makes the Request
class itself more universally usable. It also greatly simplifies
the Request handling code as well as working with the Request
class in more demanding situations.

This change makes the RequestProxy and InterRequest classes
obsolete. They will be removed in subsequent changes.